requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==41.0.7
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
python-dotenv==1.1.1
PyYAML==6.0.1
Werkzeug==3.1.3
